backend/pyproject.toml
[project]
name = "milemind-backend"
version = "0.1.0"
description = "MileMind - AI-powered running training optimizer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.49.0",
    "numpy>=2.2.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "authlib>=1.4.0",
    "python-jose[cryptography]>=3.3.0",
    "itsdangerous>=2.2.0",
    "cryptography>=44.0.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "black>=25.1.0",
    "aiosqlite>=0.21.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E402",   # module-level imports not at top (deps.py logger pattern)
    "E501",   # line too long (prompt strings, HTML templates, JSON examples)
    "UP042",  # str+Enum -> StrEnum (intentional for JSON serialization compat)
]

[tool.ruff.lint.per-file-ignores]
# Test files use PascalCase for Mock/patched-class locals and have
# legitimate reasons to diverge from PEP8 naming.
"tests/**" = ["N801", "N806", "N811", "N813"]

[tool.black]
target-version = ["py312"]
line-length = 99
